Move shared functions from Esmeralda
