+++ invisible_cities/cities/components.py
@@ -30,6 +30,7 @@ from .. evm    .event_model       import                   KrEvent
 from .. evm    .event_model       import                       Hit
 from .. evm    .event_model       import                   Cluster
 from .. evm    .event_model       import             HitCollection
+from .. evm    .event_model       import                 HitEnergy
 from .. evm    .event_model       import                    MCInfo
 from .. evm    .pmaps             import                SiPMCharge
 from .. core                      import           system_of_units as units
@@ -50,6 +51,7 @@ from .. reco                      import            peak_functions as pkf
 from .. reco                      import           pmaps_functions as pmf
 from .. reco                      import            hits_functions as hif
 from .. reco                      import             wfm_functions as wfm
+from .. reco                      import         paolina_functions as plf
 from .. reco   .xy_algorithms     import                    corona
 from .. filters.s1s2_filter       import               S12Selector
 from .. filters.s1s2_filter       import               pmap_filter
@@ -63,10 +65,14 @@ from .. io     .event_filter_io   import       event_filter_writer
 from .. io     .pmaps_io          import               pmap_writer
 from .. io     .rwf_io            import             buffer_writer
 from .. io     .mcinfo_io         import            load_mchits_df
+from .. io     .dst_io            import                 df_writer
+from .. io     .hits_io           import               hits_writer
 from .. types  .ic_types          import                        xy
 from .. types  .ic_types          import                        NN
 from .. types  .ic_types          import                       NNN
 from .. types  .ic_types          import                    minmax
+from .. types  .ic_types          import        types_dict_summary
+from .. types  .ic_types          import         types_dict_tracks
 
 NoneType = type(None)
 
@@ -1002,3 +1008,248 @@ def calculate_and_save_buffers(buffer_length    : float        ,
     # Filter out order_sensors if it is not set
     buffer_definition = pipe(*filter(None, find_signal_and_write_buffers))
     return buffer_definition
+
+
+def copy_Ec_to_Ep_hit_attribute_(hitc : HitCollection) -> HitCollection:
+    """
+    The functions copies values of Ec attributes into Ep attributes. Takes as input HitCollection and returns a copy.
+    """
+    mod_hits = []
+    for hit in hitc.hits:
+        hit = Hit(hit.npeak, Cluster(hit.Q, xy(hit.X, hit.Y), hit.var, hit.nsipm),
+                  hit.Z, hit.E, xy(hit.Xpeak, hit.Ypeak), s2_energy_c=hit.Ec, Ep=hit.Ec)
+        mod_hits.append(hit)
+    mod_hitc = HitCollection(hitc.event, hitc.time, hits=mod_hits)
+    return mod_hitc
+
+
+def make_event_summary(event_number  : int              ,
+                       topology_info : pd.DataFrame     ,
+                       paolina_hits  : HitCollection,
+                       out_of_map    : bool
+                       ) -> pd.DataFrame:
+    """
+    For a given event number, timestamp, topology info dataframe, paolina hits and kdst information returns a
+    dataframe with the whole event summary.
+
+    Parameters
+    ----------
+    event_number  : int
+    timestamp     : long int
+    topology_info : DataFrame
+        Dataframe containing track information,
+        output of track_blob_info_creator_extractor
+    paolina_hits  : HitCollection
+        Hits table passed through paolina functions,
+        output of track_blob_info_creator_extractor
+    kdst          : DataFrame
+        Kdst information read from penthesilea output
+
+
+    Returns
+    ----------
+    DataFrame containing relevant per event information.
+    """
+    es = pd.DataFrame(columns=list(types_dict_summary.keys()))
+
+    ntrks = len(topology_info.index)
+    nhits = len(paolina_hits.hits)
+
+    S2ec = sum(h.Ec for h in paolina_hits.hits)
+    S2qc = -1 #not implemented yet
+
+    pos   = [h.pos for h in paolina_hits.hits]
+    x, y, z = map(np.array, zip(*pos))
+    r = np.sqrt(x**2 + y**2)
+
+    e     = [h.Ec  for h in paolina_hits.hits]
+    ave_pos = np.average(pos, weights=e, axis=0)
+    ave_r   = np.average(r  , weights=e, axis=0)
+
+
+    list_of_vars  = [event_number, S2ec, S2qc, ntrks, nhits,
+                     *ave_pos, ave_r,
+                     min(x), min(y), min(z), min(r), max(x), max(y), max(z), max(r),
+                     out_of_map]
+
+    es.loc[0] = list_of_vars
+    #change dtype of columns to match type of variables
+    es = es.apply(lambda x : x.astype(types_dict_summary[x.name]))
+    return es
+
+
+
+def track_writer(h5out, compression='ZLIB4'):
+    """
+    For a given open table returns a writer for topology info dataframe
+    """
+    def write_tracks(df):
+        return df_writer(h5out              = h5out              ,
+                         df                 = df                 ,
+                         compression        = compression        ,
+                         group_name         = 'Tracking'         ,
+                         table_name         = 'Tracks'           ,
+                         descriptive_string = 'Track information',
+                         columns_to_index   = ['event']          )
+    return write_tracks
+
+
+def summary_writer(h5out, compression='ZLIB4'):
+    """
+    For a given open table returns a writer for summary info dataframe
+    """
+    def write_summary(df):
+        return df_writer(h5out              = h5out                      ,
+                         df                 = df                         ,
+                         compression        = compression                ,
+                         group_name         = 'Summary'                  ,
+                         table_name         = 'Events'                   ,
+                         descriptive_string = 'Event summary information',
+                         columns_to_index   = ['event']                  )
+    return write_summary
+
+
+def track_blob_info_creator_extractor(vox_size         : [float, float, float],
+                                      strict_vox_size  : bool                 ,
+                                      energy_threshold : float                ,
+                                      min_voxels       : int                  ,
+                                      blob_radius      : float                ,
+                                      max_num_hits     : int
+                                     ) -> Callable:
+    """
+    For a given paolina parameters returns a function that extract tracks / blob information from a HitCollection.
+
+    Parameters
+    ----------
+    vox_size         : [float, float, float]
+        (maximum) size of voxels for track reconstruction
+    strict_vox_size  : bool
+        if False allows per event adaptive voxel size,
+        smaller of equal thatn vox_size
+    energy_threshold : float
+        if energy of end-point voxel is smaller
+        the voxel will be dropped and energy redistributed to the neighbours
+    min_voxels       : int
+        after min_voxel number of voxels is reached no dropping will happen.
+    blob_radius      : float
+        radius of blob
+
+    Returns
+    ----------
+    A function that from a given HitCollection returns a pandas DataFrame with per track information.
+    """
+    def create_extract_track_blob_info(hitc):
+        df = pd.DataFrame(columns=list(types_dict_tracks.keys()))
+        if len(hitc.hits) > max_num_hits:
+            return df, hitc, True
+        #track_hits is a new Hitcollection object that contains hits belonging to tracks, and hits that couldnt be corrected
+        track_hitc = HitCollection(hitc.event, hitc.time)
+        out_of_map = np.any(np.isnan([h.Ep for h in hitc.hits]))
+        if out_of_map:
+            #add nan hits to track_hits, the track_id will be -1
+            track_hitc.hits.extend  ([h for h in hitc.hits if np.isnan   (h.Ep)])
+            hits_without_nan       = [h for h in hitc.hits if np.isfinite(h.Ep)]
+            #create new Hitcollection object but keep the name hitc
+            hitc      = HitCollection(hitc.event, hitc.time)
+            hitc.hits = hits_without_nan
+
+        if len(hitc.hits) > 0:
+            voxels           = plf.voxelize_hits(hitc.hits, vox_size, strict_vox_size, HitEnergy.Ep)
+            (    mod_voxels,
+             dropped_voxels) = plf.drop_end_point_voxels(voxels, energy_threshold, min_voxels)
+            tracks           = plf.make_track_graphs(mod_voxels)
+
+            for v in dropped_voxels:
+                track_hitc.hits.extend(v.hits)
+
+            vox_size_x = voxels[0].size[0]
+            vox_size_y = voxels[0].size[1]
+            vox_size_z = voxels[0].size[2]
+            del(voxels)
+            #sort tracks in energy
+            tracks     = sorted(tracks, key=plf.get_track_energy, reverse=True)
+
+            track_hits = []
+            for c, t in enumerate(tracks, 0):
+                tID = c
+                energy = plf.get_track_energy(t)
+                length = plf.length(t)
+                numb_of_hits   = len([h for vox in t.nodes() for h in vox.hits])
+                numb_of_voxels = len(t.nodes())
+                numb_of_tracks = len(tracks   )
+                pos   = [h.pos for v in t.nodes() for h in v.hits]
+                x, y, z = map(np.array, zip(*pos))
+                r = np.sqrt(x**2 + y**2)
+
+                e     = [h.Ep for v in t.nodes() for h in v.hits]
+                ave_pos = np.average(pos, weights=e, axis=0)
+                ave_r   = np.average(r  , weights=e, axis=0)
+                extr1, extr2 = plf.find_extrema(t)
+                extr1_pos = extr1.XYZ
+                extr2_pos = extr2.XYZ
+
+                blob_pos1, blob_pos2 = plf.blob_centres(t, blob_radius)
+
+                e_blob1, e_blob2, hits_blob1, hits_blob2 = plf.blob_energies_and_hits(t, blob_radius)
+                overlap = float(sum(h.Ep for h in set(hits_blob1).intersection(set(hits_blob2))))
+                list_of_vars = [hitc.event, tID, energy, length, numb_of_voxels,
+                                numb_of_hits, numb_of_tracks,
+                                min(x), min(y), min(z), min(r), max(x), max(y), max(z), max(r),
+                                *ave_pos, ave_r, *extr1_pos,
+                                *extr2_pos, *blob_pos1, *blob_pos2,
+                                e_blob1, e_blob2, overlap,
+                                vox_size_x, vox_size_y, vox_size_z]
+
+                df.loc[c] = list_of_vars
+
+                for vox in t.nodes():
+                    for hit in vox.hits:
+                        hit.track_id = tID
+                        track_hits.append(hit)
+
+            #change dtype of columns to match type of variables
+            df = df.apply(lambda x : x.astype(types_dict_tracks[x.name]))
+            track_hitc.hits.extend(track_hits)
+        return df, track_hitc, out_of_map
+
+    return create_extract_track_blob_info
+
+
+def compute_and_write_tracks_info(paolina_params, h5out):
+
+    # Create tracks and compute topology-related information
+    create_extract_track_blob_info = fl.map(track_blob_info_creator_extractor(**paolina_params),
+                                            args = 'Ep_hits',
+                                            out  = ('topology_info', 'paolina_hits', 'out_of_map'))
+
+    # Filter empty topology events
+    filter_events_topology         = fl.map(lambda x : len(x) > 0,
+                                            args = 'topology_info',
+                                            out  = 'topology_passed')
+    events_passed_topology         = fl.count_filter(bool, args="topology_passed")
+
+    # Create table with summary information
+    make_final_summary             = fl.map(make_event_summary,
+                                            args = ('event_number', 'topology_info', 'paolina_hits', 'out_of_map'),
+                                            out  = 'event_info')
+
+
+    # Define writers and make them sinks
+    write_tracks          = fl.sink(   track_writer     (h5out=h5out)             , args="topology_info"      )
+    write_summary         = fl.sink( summary_writer     (h5out=h5out)             , args="event_info"         )
+    write_topology_filter = fl.sink( event_filter_writer(h5out, "topology_select"), args=("event_number", "topology_passed"    ))
+    write_hits_paolina    = fl.sink(    hits_writer     (h5out, group_name="CHITS", table_name="highTh")
+                                                                                  , args="paolina_hits"       )
+
+
+    fn_list = (create_extract_track_blob_info              ,
+               filter_events_topology                      ,
+               fl.branch(make_final_summary, write_summary),
+               fl.branch(write_topology_filter)            ,
+               fl.branch(write_hits_paolina)               ,
+               events_passed_topology.   filter            ,
+               fl.branch(write_tracks)                    ,)
+
+    compute_tracks = pipe(*fn_list)
+
+    return compute_tracks
